+++ jwst/datamodels/__init__.py
@@ -41,9 +41,9 @@ from .nirspec_flat import NirspecFlatModel, NirspecQuadFlatModel
 from .outlierpars import OutlierParsModel
 from .pathloss import PathlossModel
 from .persat import PersistenceSatModel
-from .photom import PhotomModel, FgsPhotomModel, NircamPhotomModel, NirissPhotomModel
-from .photom import NirspecPhotomModel, NirspecFSPhotomModel
-from .photom import MiriImgPhotomModel, MiriMrsPhotomModel
+from .photom import PhotomModel, FgsPhotomModel, FgsImgPhotomModel, NircamPhotomModel, NrcImgPhotomModel
+from .photom import NirissPhotomModel, NisImgPhotomModel, NirspecPhotomModel, NirspecFSPhotomModel
+from .photom import MiriImgPhotomModel, MirImgPhotomModel, MiriMrsPhotomModel
 from .pixelarea import PixelAreaModel, NirspecSlitAreaModel, NirspecMosAreaModel, NirspecIfuAreaModel
 from .psfmask import PsfMaskModel
 from .quad import QuadModel
@@ -104,8 +104,9 @@ __all__ = [
     'PathlossModel',
     'PersistenceSatModel',
     'PixelAreaModel', 'NirspecSlitAreaModel', 'NirspecMosAreaModel', 'NirspecIfuAreaModel',
-    'PhotomModel', 'FgsPhotomModel', 'MiriImgPhotomModel', 'MiriMrsPhotomModel',
-    'NircamPhotomModel', 'NirissPhotomModel', 'NirspecPhotomModel', 'NirspecFSPhotomModel',
+    'PhotomModel', 'FgsPhotomModel', 'FgsImgPhotomModel', 'MiriImgPhotomModel', 'MirImgPhotomModel', 'MiriMrsPhotomModel',
+    'NircamPhotomModel', 'NrcImgPhotomModel', 'NirissPhotomModel', 'NisImgPhotomModel',
+    'NirspecPhotomModel', 'NirspecFSPhotomModel',
     'PsfMaskModel',
     'QuadModel', 'RampModel', 'MIRIRampModel',
     'RampFitOutputModel', 'ReadnoiseModel',